Add support for arbitrary C# `object` types as inputs (#60)
The `evaluate<T>` and `evaluateDefault<T>` methods now allow arbitrary C# objects as inputs to a policy. This makes the API much more ergonomic to use, at the expense of a bit of serdes overhead under-the-hood.

We serialize the incoming object to JSON, and then attempt to convert it to a `Dictionary<string, object>`. This is reasonable in context, because other primitive JSON types should be matched by more specific methods, like the `string` or `double` methods, for instance.

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>