Styra/Opa/OpaClient: Stricter null-handling + object deserialize fixes
This commit includes two sweeping, and closely related changes in the
OpaClient class:

 - Object deserialization to arbitrary C# types now generally works as
   expected, and can take hints from Newtonsoft.Json attribute
   annotations.
 - All methods now return *exactly* type `T`, not `T?`. If the caller
   wants a possibly nullable return type, they must explicitly provide
   the nullable version. (Ex: `evaluate<int?>(...)`)

Breaking change:
We now return OpaExceptions in cases where the library would have
previously experienced an error condition and returned a null. This is
expected to surface issues faster during development, and also has
allowed substantial cleanups in the codebase.

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>