Styra/Opa: Add evaluateBatch<T> support.
This commit adds support for a variation of the `evaluateBatch` API,
which allows returning a `Dictionary<string, T>`, instead of an
`OpaResult` type. This saves the caller a result-extraction step, and
while it's a bit opinionated (you typically can't return 2x different
types on success cases), it does make many usecases a lot cleaner.

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>