test: Add OutputConsumer settings to fixtures for better debugging.
In the OPA ASP.NET Core SDK, this setting helped reveal a syntax error
in our Rego that came up because of the OPA v1.0.0 upgrade. If we'd had
that logging in place from the beginning, it could've saved a lot of
time, so we're plumbing it in here for the sake of future maintainers.

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>