WIP: Add GetFilters high-level API for EOPA.
This commit wraps up the state of work so far on adding Enterprise OPA's
Data Filters feature, which uses the extended Compile API to generate
data filters and column masking rules from Rego policies.

Currently, we have to manually work around a codegen bug in Speakeasy,
so the low-level machinery is just manually assembling the HTTP request
the hard way.

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>